+++ Code/Common/itkLinearInterpolateImageFunction.h
@@ -48,7 +48,7 @@ namespace itk
  *
  * \ingroup ImageFunctions ImageInterpolators 
  */
-template <class TInputImage, class TCoordRep = float>
+template <class TInputImage, class TCoordRep = double>
 class ITK_EXPORT LinearInterpolateImageFunction : 
   public InterpolateImageFunction<TInputImage,TCoordRep> 
 {
