+++ Code/Common/itkVectorLinearInterpolateImageFunction.h
@@ -38,7 +38,7 @@ namespace itk
  * \ingroup ImageFunctions ImageInterpolators
  * 
  */
-template <class TInputImage, class TCoordRep = float>
+template <class TInputImage, class TCoordRep = double>
 class ITK_EXPORT VectorLinearInterpolateImageFunction : 
   public VectorInterpolateImageFunction<TInputImage,TCoordRep> 
 {
